WORK IN PROGRESS: Terrain generation by itself
What im trying to do now is make it so when you are halfway through the game object that checks for when you hit the collider, generate a new path, but im stuck and too sleepy to fix it. I also can't call generate new terrain so i need to look at this again tommorrow.